create form valid

+++ frontend/cmd/web/models.go
@@ -215,7 +215,7 @@ type CreateVolume struct {
 	VolumeName      string `json:"volume_name,omitempty"`
 	VolumeSize      uint64 `json:"volume_size,omitempty"`
 	VolumeIOProfile string `json:"volume_io_profile,omitempty"`
-	VolumeHALevel   uint32 `json:"volume_ha_level,omitempty"`
+	VolumeHALevel   uint64 `json:"volume_ha_level,omitempty"`
 	VolumeEncrypted bool   `json:"volume_encrypted,omitempty"`
 	VolumeSharedv4  bool   `json:"volume_sharedv4,omitempty"`
 	VolumeNoDiscard bool   `json:"volume_no_discard,omitempty"`
